added interface to HypothesisTests.jl and added some docs

+++ src/ranks.jl
@@ -1,7 +1,7 @@
 function ranks(v::AbstractArray)
   n     = length(v)
   place = sortperm(v)
-  ord   = Array(Float64, n)
+  ord   = Array(Int64, n)
 
   for(i in 1:n)
     ord[place[i]] = i
